Pass signup intent from querystring to onboarding step 1 as a hidden field.
