+++ frontend/lib/pages/login-page.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import querystring from 'querystring';
 
 import Page from '../page';
 import Routes, { routeMap } from '../routes';
@@ -15,6 +14,7 @@ import hardRedirect from '../tests/hard-redirect';
 import { PhoneNumberFormField } from '../phone-number-form-field';
 import { assertNotNull } from '../util';
 import { OutboundLink } from '../google-analytics';
+import { getPostOrQuerystringVar } from '../querystring';
 
 const NEXT = 'next';
 
@@ -71,52 +71,6 @@ export class LoginForm extends React.Component<LoginFormProps> {
       </SessionUpdatingFormSubmitter>
     );
   }
-}
-
-/**
- * This is intentionally structured as a subset of react-router's
- * router context, to make it easy to interoperate with.
- */
-type LocationSearchInfo = {
-  location: {
-    search: string
-  }
-};
-
-/**
- * Return the value of the last-defined key in the given querystring.
- */
-export function getQuerystringVar(routeInfo: LocationSearchInfo, name: string): string|undefined {
-  let val = querystring.parse(routeInfo.location.search.slice(1))[name];
-
-  if (Array.isArray(val)) {
-    val = val[val.length - 1];
-  }
-
-  return val;
-}
-
-/**
- * This is intentionally structured as a subset of our app context,
- * to make it easy to interoperate with.
- */
-type HttpPostInfo = {
-  legacyFormSubmission?: {
-    POST: Partial<{ [key: string]: string }>;
-  }
-};
-
-/**
- * If this is a POST, returns the last value of the given key, or undefined if
- * not present.
- *
- * Otherwise, returns the last-defined key in the given querystring.
- */
-export function getPostOrQuerystringVar(info: LocationSearchInfo & HttpPostInfo, name: string): string|undefined {
-  if (info.legacyFormSubmission) {
-    return info.legacyFormSubmission.POST[name];
-  }
-  return getQuerystringVar(info, name);
 }
 
 /**
