+++ frontend/lib/tests/pages/login-page.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoginPage, { getQuerystringVar, getPostOrQuerystringVar, performHardOrSoftRedirect, absolutifyURLToOurOrigin } from '../../pages/login-page';
+import LoginPage, { performHardOrSoftRedirect, absolutifyURLToOurOrigin } from '../../pages/login-page';
 import { Route } from 'react-router';
 import { AppTesterPal } from '../app-tester-pal';
 import { setHardRedirector } from '../hard-redirect';
@@ -11,45 +11,6 @@ test('login page sets "next" input to expected value', () => {
   });
   pal.rr.getByText(/Sign in/i);
   expect(pal.getElement('input', '[name="next"]').value).toEqual('https://blarg.com/bop');
-});
-
-const searchInfo = (search: string) => ({
-  location: { search }
-});
-
-const postInfo = (POST: Partial<{ [key: string]: string }>, search = '') => ({
-  location: { search },
-  legacyFormSubmission: { POST },
-});
-
-describe('getQuerystringVar()', () => {
-  it('returns the only value when it exists', () => {
-    expect(getQuerystringVar(searchInfo('?foo=bar'), 'foo')).toBe('bar');
-    expect(getQuerystringVar(searchInfo('?foo='), 'foo')).toBe('');
-  });
-
-  it('returns the last value when multiple definitions exist', () => {
-    expect(getQuerystringVar(searchInfo('?foo=bar&foo=baz'), 'foo')).toBe('baz');
-  });
-
-  it('returns undefined when it does not exist', () => {
-    expect(getQuerystringVar(searchInfo(''), 'foo')).toBeUndefined();
-    expect(getQuerystringVar(searchInfo('?other=thing'), 'foo')).toBeUndefined();
-  });
-});
-
-describe('getPostOrQuerystringVar()', () => {
-  it('returns POST data when available', () => {
-    expect(getPostOrQuerystringVar(postInfo({ blah: 'oof' }), 'blah')).toBe('oof');
-  });
-
-  it('returns querystring data when POST is not available', () => {
-    expect(getPostOrQuerystringVar(searchInfo('?blah=oof'), 'blah')).toBe('oof');
-  });
-
-  it('returns undefined when POST exists but does not contain variable', () => {
-    expect(getPostOrQuerystringVar(postInfo({ blah: 'oof' }, '?zoof=z'), 'zoof')).toBeUndefined();
-  });
 });
 
 describe('performHardOrSoftRedirect()', () => {
