+++ frontend/lib/tests/signup-intent.test.ts
@@ -1,5 +1,5 @@
 import { validateDjangoChoices } from "../common-data";
-import { SignupIntentChoice } from "../signup-intent";
+import { SignupIntentChoice, validateSignupIntent, DEFAULT_SIGNUP_INTENT_CHOICE } from "../signup-intent";
 
 type SignupIntentDjangoChoice = [SignupIntentChoice, string];
 
@@ -9,4 +9,11 @@ test('SignupIntentChoice has valid choices', () => {
   for (let choice in SignupIntentChoice) {
     validateDjangoChoices(SIGNUP_INTENT_CHOICES, [choice, SignupIntentChoice[choice]]);
   }
+});
+
+test('validateSignupIntent() works', () => {
+  expect(validateSignupIntent(undefined)).toStrictEqual(DEFAULT_SIGNUP_INTENT_CHOICE);
+  expect(validateSignupIntent('blargh')).toStrictEqual(DEFAULT_SIGNUP_INTENT_CHOICE);
+  expect(validateSignupIntent('hp')).toStrictEqual(SignupIntentChoice.HP);
+  expect(validateSignupIntent('HP')).toStrictEqual(SignupIntentChoice.HP);
 });
